private topics assignment completed

+++ app/controllers/favorites_controller.rb
@@ -2,6 +2,7 @@ class FavoritesController < ApplicationController
   def create
     post = Post.find(params[:post_id])
     favorite = current_user.favorites.build(post: post)
+    authorize favorite
     if favorite.save
       flash[:notice] = "Favorite was saved."
       redirect_to [post.topic, post]
@@ -14,6 +15,7 @@ class FavoritesController < ApplicationController
   def destroy
     post = Post.find(params[:post_id])
     favorite = Favorite.find(params[:id])
+    authorize favorite
     if favorite.destroy
       flash[:notice] = "Favorite was removed."
       redirect_to [post.topic, post]
